requirements.txt
asgiref==3.8.1
Django==5.1
django-dirtyfields==1.9.3
django-rest-framework==0.1.0
django-versatileimagefield==3.1
djangorestframework==3.15.2
pillow==10.4.0
python-magic==0.4.27
sqlparse==0.5.1
typing_extensions==4.12.2
